# src/multi-chat/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kuwa API",
        "description": "API definition for KuwaClient service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/read/cloud/{path}": {
            "get": {
                "tags": [
                    "Cloud"
                ],
                "summary": "List cloud directory or file",
                "operationId": "207dd29fc51284b704c88c97a636fe2f",
                "parameters": [
                    {
                        "name": "path",
                        "in": "path",
                        "description": "The cloud directory or file path. If not provided, defaults to a single dot.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cloud data listed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/delete/cloud/{path}": {
            "delete": {
                "tags": [
                    "Cloud"
                ],
                "summary": "Delete cloud file or folder",
                "operationId": "b81ad367dc284f3ac77aef58297ad717",
                "parameters": [
                    {
                        "name": "path",
                        "in": "path",
                        "description": "Path to cloud item to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cloud item deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}